:sparkles:feat:Set Entities and basic repositories

+++ src/main/java/excluz/excluz/domain/event/event/repository/EventRepository.java
@@ -3,5 +3,6 @@ package excluz.excluz.domain.event.event.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import excluz.excluz.common.entity.Event;
 
-public interface EventRepository extends JpaRepository<Event, Long> {
+public interface EventRepository extends JpaRepository<Event, Integer> {
+
 }
