+++ src/main/java/excluz/excluz/domain/event/eventApplicant/repository/EventApplicantRepository.java
@@ -3,5 +3,5 @@ package excluz.excluz.domain.event.eventApplicant.repository;
 import excluz.excluz.common.entity.EventApplicant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventApplicantRepository extends JpaRepository<EventApplicant, Long> {
+public interface EventApplicantRepository extends JpaRepository<EventApplicant, Integer> {
 }
