+++ src/main/java/excluz/excluz/domain/event/eventItem/repository/EventItemRepository.java
@@ -3,5 +3,5 @@ package excluz.excluz.domain.event.eventItem.repository;
 import excluz.excluz.common.entity.EventItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventItemRepository extends JpaRepository<EventItem, Long> {
+public interface EventItemRepository extends JpaRepository<EventItem, Integer> {
 }
